cambie un tipo de dato en frozen_tunnel y agregue las relaciones de frozen_tunnel con subprocess y subprocess_io

+++ db/migrate/20130627062216_create_frozen_tunnels.rb
@@ -3,8 +3,8 @@ class CreateFrozenTunnels < ActiveRecord::Migration
     create_table :frozen_tunnels do |t|
       t.integer :order_number
       t.integer :tunnel_id
-      t.double :tunnel_temperature
-      t.double :packing_pallet_temperature
+      t.float :tunnel_temperature
+      t.float :packing_pallet_temperature
 
       t.timestamps
     end
